# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entity.Product:
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      unit:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: sdesain25@gmail.com
    name: API Support
    url: https://uqonstd.xyz/
  description: A documentation how to access the api's routes on your laundry application
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Laundry API Documentation
  version: "1.0"
paths:
...
  /products:
    get:
      description: Retrieve a list of all products
      parameters:
      - description: Product name
        in: query
        name: name
        type: string
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Get all products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Create a new product
      parameters:
      - description: Product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/entity.Product'
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      security:
      - ApiKeyAuth: []
      summary: Create product
      tags:
      - Products
  /products/{id}:
    delete:
      description: Delete a product
      parameters:
      - description: Product id
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      security:
      - ApiKeyAuth: []
      summary: Delete product
      tags:
      - Products
    get:
      description: Retrieve a product by id
      parameters:
      - description: Product id
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      security:
      - ApiKeyAuth: []
      summary: Get product by id
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update a product
      parameters:
      - description: Product id
        in: path
        name: id
        required: true
        type: integer
      - description: Product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/entity.Product'
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      security:
      - ApiKeyAuth: []
      summary: Update product
      tags:
      - Products
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"